requirements.txt
altair==5.5.0
asttokens==3.0.0
attrs==25.1.0
bcrypt==4.3.0
blinker==1.9.0
cachetools==5.5.2
captcha==0.7.1
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
colorlover==0.3.0
comm==0.2.2
contourpy==1.3.1
cryptography==44.0.2
cufflinks==0.17.3
cycler==0.12.1
DateTime==5.5
decorator==5.2.1
et_xmlfile==2.0.0
executing==2.2.0
extra-streamlit-components==0.1.80
fonttools==4.55.3
gitdb==4.0.12
GitPython==3.1.44
graphviz==0.20.3
greenlet==3.1.1
idna==3.10
ipython==9.1.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.6
jedi==0.19.2
Jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyterlab_widgets==3.0.14
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.0
matplotlib-inline==0.1.7
narwhals==1.30.0
numpy==2.2.1
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
parso==0.8.4
patsy==1.0.1
pi==0.1.2
pillow==11.1.0
plotly==6.0.1
plotly-express==0.4.1
prompt_toolkit==3.0.51
protobuf==5.29.3
# psycopg2==2.9.10
psycopg2-binary==2.9.10
pure_eval==0.2.3
pyarrow==19.0.1
pycparser==2.22
pydeck==0.9.1
Pygments==2.19.1
PyJWT==2.10.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
referencing==0.36.2
reportlab==4.3.1
requests==2.32.3
rpds-py==0.23.1
scipy==1.15.2
seaborn==0.13.2
setuptools==76.0.0
six==1.17.0
smmap==5.0.2
SQLAlchemy==2.0.39
stack-data==0.6.3
statsmodels==0.14.4
streamlit==1.43.1
streamlit-authenticator==0.4.2
streamlit-autorefresh==1.0.1
streamlit-elements==0.1.0
streamlit-option-menu==0.4.0
tenacity==9.0.0
toml==0.10.2
tornado==6.4.2
traitlets==5.14.3
typing_extensions==4.12.2
tz==0.2.2
tzdata==2024.2
urllib3==2.3.0
utils==1.0.2
watchdog==6.0.0
wcwidth==0.2.13
wheel==0.45.1
widgetsnbextension==4.0.14
XlsxWriter==3.2.2
zope.interface==7.2
